docs/compare-mongodb-vs-postgres/research/pyproject.toml
[project]
name = "compare-postgres-mongo-service"
version = "0.1.0"
description = "Research for comparing mongodb and postgres microservice"
requires-python = ">=3.12,<4.0"
dependencies = [
    "asyncpg>=0.30.0,<1.0.0",
    "psycopg2>=2.9.10,<3.0.0",
    "psycopg>=3.2.9",
    "motor>=3.7.1,<3.8.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "pydantic-settings>=2.8.1,<3.0.0",
    "greenlet>=3.2.0,<4.0.0",
    "aiohttp (>=3.11.18,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
